Cleanup of old fixes
The script is applying changes that the spec has already fixed in the
latest versions.

Additionally, there are some "fixes" that actually break the current
schema (and our code), like `{Properties}{Tags}{ItemType}= 'Tag'`, which
doesn't seem to be valid anymore.

This deletes all the old fixes to the data structure, and adds the one
that is currently failing.